backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
apscheduler
pymysql
python-dotenv
requests
psycopg2-binary